POC for JWS-PS256 signing with HSM (hardware-encryption-device)
Proof-of-concept for performing JWS-PS256 (RSA-PSS-MGF1-SHA256) signing
on hardware-encryption-devices that may not support the transformation
scheme but support Plain-RSA-Signing/Encryption never the less.

The hashing and padding is performed at the application layer and then
the resulting byte is then input to rsa-encryption mechanism,
effectively simulating a signing using rsa-pss-sha256-with-mgf1-sha256
scheme.

Bouncy-Castle's 'RSAPSSSigner' has been modified to obtain the desired
result.

Initial commit with test cases!

Includes readme notes for overview.